+++ streams/functions/src/logs-processor/LogDocumentId.ts
@@ -1,6 +1,6 @@
 import { ethers } from 'ethers';
 
-export class LogRowId {
+export class LogDocumentId {
   public static create(chainId: number, transactionHash: string, logIndex: string): string {
     const safeTransactionHash = transactionHash.toLowerCase();
     const rawId = ethers.utils.toUtf8Bytes(`${chainId};${safeTransactionHash};${logIndex}`);