+++ streams/functions/src/logs-processor/LogsProcessor.ts
@@ -2,31 +2,27 @@ import { IWebhook } from '@moralisweb3/streams-typings';
 
 import { CollectionNameBuilder } from '../core/CollectionNameBuilder';
 import { LogParser } from './LogParser';
-import { LogRow, LogRowBuilder } from './LogRowBuilder';
+import { LogDocument, LogDocumentBuilder } from './LogDocumentBuilder';
 
 export class LogsProcessor {
   private readonly collectionNameBuilder = new CollectionNameBuilder();
 
-  public process(batch: IWebhook): LogRowUpdate[] {
-    const updates: LogRowUpdate[] = [];
+  public process(batch: IWebhook): LogDocumentUpdate[] {
+    const updates: LogDocumentUpdate[] = [];
 
-    const logParsers: Record<string, LogParser> = {};
-    for (const streamId of Object.keys(batch.abis)) {
-      logParsers[streamId] = new LogParser(batch.abis[streamId]);
+    if (batch.abi.length < 1) {
+      return updates;
     }
 
-    for (const log of batch.logs) {
-      const logParser = logParsers[log.streamId];
-      if (!logParser) {
-        continue;
-      }
+    const logParser = new LogParser(batch.abi);
 
+    for (const log of batch.logs) {
       const logParams = logParser.read(log);
-      const row = LogRowBuilder.build(log, logParams, batch.block, batch.confirmed, batch.chainId);
+      const document = LogDocumentBuilder.build(log, logParams, batch.block, batch.confirmed, batch.chainId);
 
       updates.push({
-        tableName: this.collectionNameBuilder.build(log.tag),
-        row,
+        collectionName: this.collectionNameBuilder.build(batch.tag),
+        document,
       });
     }
 
@@ -34,7 +30,7 @@ export class LogsProcessor {
   }
 }
 
-export interface LogRowUpdate {
-  tableName: string;
-  row: LogRow;
+export interface LogDocumentUpdate {
+  collectionName: string;
+  document: LogDocument;
 }