+++ streams/functions/src/store/TxsWriter.ts
@@ -1,24 +1,24 @@
 import * as admin from 'firebase-admin';
 
-import { TxRow } from '../txs-processor/TxRowBuilder';
-import { TxRowUpdate } from '../txs-processor/TxsProcessor';
+import { TxDocument } from '../txs-processor/TxDocumentBuilder';
+import { TxDocumentUpdate } from '../txs-processor/TxsProcessor';
 
 export class TxsWriter {
   public constructor(private readonly store: admin.firestore.Firestore) {}
 
-  public writeMany(updates: TxRowUpdate[]): Promise<void>[] {
+  public writeMany(updates: TxDocumentUpdate[]): Promise<void>[] {
     return updates.map((update) => this.store.runTransaction(() => this.write(update)));
   }
 
-  private async write(update: TxRowUpdate): Promise<void> {
+  private async write(update: TxDocumentUpdate): Promise<void> {
     const collection = this.store.collection(`moralis/txs/${update.collectionName}`);
 
-    const itemDoc = collection.doc(update.row.id);
+    const itemDoc = collection.doc(update.document.id);
 
-    if (!update.row.confirmed) {
+    if (!update.document.confirmed) {
       const item = await itemDoc.get();
       if (item.exists) {
-        const log = item.data() as TxRow;
+        const log = item.data() as TxDocument;
         if (log.confirmed) {
           // We don't update a confirmed item with a non confirmed item.
           return;
@@ -27,7 +27,7 @@ export class TxsWriter {
     }
 
     await itemDoc.set({
-      ...update.row,
+      ...update.document,
       updatedAt: admin.firestore.FieldValue.serverTimestamp(),
     });
   }